# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Messaging app API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OkResponse {
  ok: boolean;
}

export interface ErrorResponse {
  error: string;
}

export interface User {
  id: number;
  username: string;
  displayName: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  bio?: string | null;
  isOnline: boolean;
  lastSeen: string;
  createdAt: string;
}

export interface AuthUser {
  id: number;
  username: string;
  displayName: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  bio?: string | null;
  isOnline: boolean;
  lastSeen: string;
  createdAt: string;
}

export interface RegisterInput {
  /**
     * @minLength 3
     * @maxLength 30
     */
  username: string;
  /**
     * @minLength 1
     * @maxLength 50
     */
  displayName: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export interface UserUpdate {
  /**
     * @minLength 1
     * @maxLength 50
     */
  displayName?: string;
  /** @maxLength 200 */
  bio?: string;
  avatarUrl?: string;
}

export interface Reaction {
  emoji: string;
  count: number;
  userIds: number[];
}

export interface Message {
  id: number;
  conversationId: number;
  senderId: number;
  sender?: User | null;
  content: string;
  /** @nullable */
  editedAt?: string | null;
  createdAt: string;
  reactions: Reaction[];
}

export interface Conversation {
  id: number;
  participants: User[];
  lastMessage?: Message | null;
  unreadCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface ConversationInput {
  participantId: number;
}

export interface MessageInput {
  /** @minLength 1 */
  content: string;
}

export interface MessageUpdate {
  /** @minLength 1 */
  content: string;
}

export interface ReactionInput {
  emoji: string;
}

export interface TypingInput {
  isTyping: boolean;
}

export interface TypingUser {
  userId: number;
  username: string;
  displayName: string;
}

export interface MessageSearchResult {
  message: Message;
  conversationId: number;
}

export type ListUsersParams = {
search?: string;
};

export type GetMessagesParams = {
/**
 * Load messages before this message ID (pagination)
 */
before?: number;
limit?: number;
};

export type SearchMessagesParams = {
q: string;
};