# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Messaging app API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: users
    description: User management
  - name: conversations
    description: Conversations
  - name: messages
    description: Messages
  - name: reactions
    description: Message reactions

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Log out
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List or search users
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"

  /users/me:
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /users/me/heartbeat:
    post:
      operationId: heartbeat
      tags: [users]
      summary: Mark current user as online
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"

  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /conversations:
    get:
      operationId: listConversations
      tags: [conversations]
      summary: List conversations for the current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
    post:
      operationId: createConversation
      tags: [conversations]
      summary: Create or get a conversation with another user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConversationInput"
      responses:
        "200":
          description: OK (existing conversation returned)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"

  /conversations/{conversationId}:
    get:
      operationId: getConversation
      tags: [conversations]
      summary: Get a single conversation
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /conversations/{conversationId}/read:
    post:
      operationId: markRead
      tags: [conversations]
      summary: Mark all messages in a conversation as read
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"

  /conversations/{conversationId}/typing:
    post:
      operationId: setTyping
      tags: [conversations]
      summary: Set typing indicator
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TypingInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"

  /conversations/{conversationId}/typing-status:
    get:
      operationId: getTypingStatus
      tags: [conversations]
      summary: Get who is currently typing in a conversation
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TypingUser"

  /conversations/{conversationId}/messages:
    get:
      operationId: getMessages
      tags: [messages]
      summary: List messages in a conversation
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
        - name: before
          in: query
          required: false
          schema:
            type: integer
          description: Load messages before this message ID (pagination)
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  /conversations/{conversationId}/messages/{messageId}:
    patch:
      operationId: editMessage
      tags: [messages]
      summary: Edit a message
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
    delete:
      operationId: deleteMessage
      tags: [messages]
      summary: Delete a message
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"

  /conversations/{conversationId}/messages/{messageId}/reactions:
    post:
      operationId: addReaction
      tags: [reactions]
      summary: Add an emoji reaction to a message
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReactionInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  /conversations/{conversationId}/messages/{messageId}/reactions/{emoji}:
    delete:
      operationId: removeReaction
      tags: [reactions]
      summary: Remove an emoji reaction from a message
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
        - name: emoji
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  /messages/search:
    get:
      operationId: searchMessages
      tags: [messages]
      summary: Search messages across all conversations
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MessageSearchResult"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
      required:
        - ok

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    User:
      type: object
      required: [id, username, displayName, isOnline, lastSeen, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        isOnline:
          type: boolean
        lastSeen:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time

    AuthUser:
      type: object
      required: [id, username, displayName, isOnline, lastSeen, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        isOnline:
          type: boolean
        lastSeen:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time

    RegisterInput:
      type: object
      required: [username, displayName, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 30
        displayName:
          type: string
          minLength: 1
          maxLength: 50
        password:
          type: string
          minLength: 6

    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string

    UserUpdate:
      type: object
      properties:
        displayName:
          type: string
          minLength: 1
          maxLength: 50
        bio:
          type: string
          maxLength: 200
        avatarUrl:
          type: string

    Conversation:
      type: object
      required: [id, participants, unreadCount, createdAt, updatedAt]
      properties:
        id:
          type: integer
        participants:
          type: array
          items:
            $ref: "#/components/schemas/User"
        lastMessage:
          anyOf:
            - $ref: "#/components/schemas/Message"
            - type: "null"
        unreadCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    ConversationInput:
      type: object
      required: [participantId]
      properties:
        participantId:
          type: integer

    Message:
      type: object
      required: [id, conversationId, senderId, content, createdAt, reactions]
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        senderId:
          type: integer
        sender:
          anyOf:
            - $ref: "#/components/schemas/User"
            - type: "null"
        content:
          type: string
        editedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        reactions:
          type: array
          items:
            $ref: "#/components/schemas/Reaction"

    Reaction:
      type: object
      required: [emoji, count, userIds]
      properties:
        emoji:
          type: string
        count:
          type: integer
        userIds:
          type: array
          items:
            type: integer

    MessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1

    MessageUpdate:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1

    ReactionInput:
      type: object
      required: [emoji]
      properties:
        emoji:
          type: string

    TypingInput:
      type: object
      required: [isTyping]
      properties:
        isTyping:
          type: boolean

    TypingUser:
      type: object
      required: [userId, username, displayName]
      properties:
        userId:
          type: integer
        username:
          type: string
        displayName:
          type: string

    MessageSearchResult:
      type: object
      required: [message, conversationId]
      properties:
        message:
          $ref: "#/components/schemas/Message"
        conversationId:
          type: integer